+++ tests/testthat/test-matrix.R
@@ -27,7 +27,7 @@ expect_identical(getRatings(r[c("u1", "u2"), colnames(r)]),
     colnames(r)]))
 
 ## subset
-expect_identical(db[2:5,2:5], as(r[2:5,2:5], "matrix"))
+expect_identical(db[9:10,2:5], as(r[9:10,2:5], "matrix"))
 
 ## number of ratings
 expect_identical(sum(sapply(as(r, "list"), length)), nrow(as(r, "data.frame")))
@@ -35,4 +35,9 @@ expect_identical(nratings(r), sum(sapply(as(r, "list"), length)))
 
 ## check ratings
 expect_identical(as(r, "dgTMatrix")@x, getRatings(r))
+
+## check subset assignment (preserves 0s)
+r[2,1:5] <- 1:5
+db[2,1:5] <- 1:5
+expect_identical(as(r, "matrix"), db)